requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl==3.1.2

# Machine Learning
scikit-learn==1.4.0
scipy==1.12.0

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
PyPDF2==3.0.1

# AI Integration
anthropic==0.25.0
python-dotenv==1.0.0
mistralai>=0.4.0

# Garmin Connect Integration
garminconnect==0.2.13

# Google Calendar Integration
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database (optional - for storing processed data)
sqlalchemy==2.0.25

# Utilities
python-dateutil==2.8.2
watchdog==4.0.0  # File system monitoring for auto-import
requests==2.31.0  # HTTP requests for API calls
